+++ src/main/java/domain/connection/ConnectionGenerator.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.connection;
 
 public interface ConnectionGenerator {
