+++ src/main/java/domain/prize/Prize.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.prize;
+
+import domain.ColumnPosition;
 
 public class Prize {
