+++ src/main/java/domain/prize/PrizeName.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.prize;
 
 public class PrizeName {
