+++ src/main/java/domain/prize/Prizes.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.prize;
 
+import domain.ColumnPosition;
 import java.util.List;
 
 public class Prizes {
