+++ src/main/java/domain/result/LadderResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public class LadderResult {
