+++ src/test/java/domain/connection/NonContinuousConnectionGeneratorTest.java
@@ -1,8 +1,8 @@
-package domain;
+package domain.connection;
 
-import static domain.Connection.DISCONNECTION;
-import static domain.Connection.LEFT_CONNECTION;
-import static domain.Connection.RIGHT_CONNECTION;
+import static domain.connection.Connection.DISCONNECTION;
+import static domain.connection.Connection.LEFT_CONNECTION;
+import static domain.connection.Connection.RIGHT_CONNECTION;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.BeforeEach;
