+++ src/test/java/domain/line/PointTest.java
@@ -1,10 +1,11 @@
-package domain;
+package domain.line;
 
-import static domain.Connection.DISCONNECTION;
-import static domain.Connection.LEFT_CONNECTION;
-import static domain.Connection.RIGHT_CONNECTION;
+import static domain.connection.Connection.DISCONNECTION;
+import static domain.connection.Connection.LEFT_CONNECTION;
+import static domain.connection.Connection.RIGHT_CONNECTION;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.ColumnPosition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
