+++ src/test/java/domain/prize/PrizeNameTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.prize;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
